+++ class/animales_voladores.h
@@ -66,7 +66,7 @@ class Cerdito:public Animal_volador
 	public:
 
 	virtual unsigned int obtener_recurso() const {return 1;}
-	virtual unsigned int obtener_alpha() const {return 255;}
+	virtual Uint8 obtener_alpha() const {return 255;}
 
 	virtual void configurar_posicion(SDL_Rect& r) const
 	{
@@ -106,7 +106,7 @@ class Vaquita:public Animal_volador
 	public:
 
 	virtual unsigned int obtener_recurso() const {return 1;}
-	virtual unsigned int obtener_alpha() const {return 255;}
+	virtual Uint8 obtener_alpha() const {return 255;}
 
 	virtual void configurar_posicion(SDL_Rect& r) const
 	{
