+++ class/bonus.h
@@ -34,7 +34,7 @@ class Bonus:public Actor_movil, public Representable
 	public:
 
 	virtual unsigned int obtener_recurso() const {return 1;}
-	virtual unsigned int obtener_alpha() const {return 255;}
+	virtual Uint8 obtener_alpha() const {return 255;}
 
 	virtual void configurar_posicion(SDL_Rect& r) const
 	{
